+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [v0.4.3-dev] - Unreleased
 
 ### Added
--
+- Added loglevel 4 bulk diagnostic mode to emit verbose Scryfall bulk row diagnostics.
 
 ### Changed
 -